convenience helpers for creating captured services
captured services are processes that run indefinitely, but print
backtraces to stdout.
For that purpose, the command object needs to be supplied with a
'compiler' argument, which is passed on to the tmux pane when the
command executes.
When parsing the pane's output, the existence of the 'compiler'
attribute now temporarily forces the corresponding compiler settings,
making it possible to read the service's backtrace.